Remove redundant settings screen class
